+++ CHANGELOG.md
@@ -8,7 +8,9 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 
 ### Added
 * add type generation ([#3830][] @voxpelli)
+* [`no-unescaped-entities`]: add suggestions ([#3831][] @StyleShit)
 
+[#3831]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3831
 [#3830]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3830
 
 ## [7.36.1] - 2024.09.12